test: move spin polarization restart test to python interface
